float > long 변환

+++ Assets/WOM/Scripts/Common/SkilSlot.cs
@@ -60,7 +60,7 @@ public class SkilSlot : MonoBehaviour
 
     public void SetTxt_Cost(string value)
     {
-        var txtSymbolValue = UtilityMethod.ChangeSymbolNumber(float.Parse(value));
+        var txtSymbolValue = UtilityMethod.ChangeSymbolNumber(long.Parse(value));
         txtCost.text = txtSymbolValue;
     }
